fix: Remove unused identifier from NviStatusRequest (#442)
Removes the candidateId from the request, as this is taken from path parameters and the update request applied directly to the Candidate. A previous PR added validation on this field, which causes errors as the field is not actually in use.

Also adds an empty allowedOperations field to the now-deprecated builder, to ensure the field is always present (as the OpenAPI spec promises).