[np-48561] refactor: Use identity service in UpsertAssigneeHandler (#444)
Ticket: https://sikt.atlassian.net/browse/NP-48561

This PR refactors access control in UpsertAssigneeHandler to use identityServiceClient instead of manually performing the same call. It is done in conjunction with this ticket to simplify testing of other necessary changes in the handler.

The authorization check in this handler is slightly more complex because we check the access level of the user being assigned and not just the user making the request.