[np-49000] fix: Improve parsing and test setup (#475)
Ticket: https://sikt.atlassian.net/browse/NP-49000


Makes more fields optional during the SPARQL parsing flow, to allow
handling of drafts and other documents with incomplete data. Changes
validation logic in the new DTOs to only check that required fields are
present, and not that the content is valid. This change to validation
aims to broadly split the work in three parts:

1. Parsing via SPARQL and deserialization with Jackson: Minimal
validation, only require fields that cannot be absent in a document
(i.e. publication ID). Errors here are unexpected and unhandled (thrown
to DLQ for debugging).
2. Validation of DTO: Check that all expected fields are set, after
`EvaluatorService` has determined that the publication should be
evaluated. Errors here result in the publication being evaluated as a
`NonCandidate`.
3. Domain validation in `EvaluatorService`, `PointService`, and
`CreatorVerificationUtil`: Check for valid values in relevant fields.
This should be consolidated and reviewed, but the tests still pass.
Minor issues here may affect evaluation (change in points or
`NonCandidate` status). Fatal errors here are unexpected and unhandled
(thrown to DLQ for debugging).