[np-48456] feat: Expand and refactor Candidate model (#484)
Ticket: https://sikt.atlassian.net/browse/NP-48456


This is a major rewrite of the data model in order to resolve several of
the sub-tickets on this issue. The goal is to persist various
publication metadata that we need and to add this data to existing
records via a migration (using source data in S3).

Core changes:

- Significant model refactoring that tries to use the same pattern
throughout the project (`DTO` <-> `Model` <-> `DAO`) . This isn't
completely finished, but I've tried to follow the pattern in the classes
that needed to be rewritten/expanded anyway.
- Persists relevant metadata, including the names of NVI-affiliated
creators, all top-level organizations affiliated with the publication,
page count, and abstract text.
- Splits the `Candidate`/`CandidateDao.DbCandidate` model into two main
sub-classes (`PublicationDetails` and `PointCalculation`) in order to
reduce the amount of code and number of fields in each top-level class.
- Rewrites many existing tests, particularly in
`EvaluateNviCandidateHandlerTest`:
- These now largely use the new `SampleExpandedPublicationFactory` setup
to generate dynamic data.
- Many test cases have been rewritten to assert specific conditions
relevant to the test, rather than generating the whole expected output
document and comparing equality.