feat: Add integration testing (#485)
Changes:
- Adds a new Gradle subproject for integration tests with a minimal
example of Cucumber tests covering multiple handlers. This is necessary
setup for future changes and will allow us to merge many existing tests.
- Adds a new `FakeEnvironment` utility to set up fake environment
variables for tests. This is a "proof of concept" and not fully used
yet, but the goal is to remove all environment variables defined in
top-level `build.gradle` files etc.
- Includes some refactoring/clean-up of test util methods.