[np-49126] fix: Handle large documents in event flow (#486)
Ticket: https://sikt.atlassian.net/browse/NP-49126

This is (hopefully) a fix for the issue of processing large documents in
events, but we will need to run the code in `test` on the full data set
to verify.

Changes:

- Changes the persisted data model to only include organizations
relevant for NVI and not the full set of top-level organizations.
- Adds migration code that handles both Candidates with no persisted
top-level organizations and those with irrelevant organizations, by
renaming the field and creating a new, separate
`topLevelNviOrganizations` field.
- Includes the `language` metadata in the same migration, as this was
missed in the previous round.


Caveats:

- The tests are quite ugly and verbose, but they are temporary.
- This does not include tests for the actual problematic part of the
code (`DynamoDbEventToQueueHandler` and `DataEntryUpdateHandler`), but
that's a separate discussion...