[np-49176] fix: Simplify data in update events (#488)
Ticket: https://sikt.atlassian.net/browse/NP-49176

This simplifies the data being sent in the indexing event flow in order
to avoid issues with large documents on SQS queues. The actual
functionality should be unchanged, but the PR touches on many different
parts of the system and test setup.

Main points:

- A new `DynamoDbChangeMessage` record is added that only contains the
information relevant for the indexing process. The stream events from
DynamoDB are converted to this immediately in
`DynamoDbEventToQueueHandler` and the new model passed through the rest
of the event flow.
- Some existing test suites had to be rewritten entirely, and I took the
opportunity to do this as integration tests (testing both handlers in
the same suite). This will make merging the handlers easier later on.
- Some existing test cases were deleted, as the scenarios tested were no
longer possible.