[np-49197] fix: Handle contributors with multiple roles (#491)
Ticket: https://sikt.atlassian.net/browse/NP-49197

Adds handling of documents where the same contributors is listed
multiple times with different roles.
For simplicity, this is done by discarding all contributors that are not
creators, as they are irrelevant here.
If non-creator contributors become relevant, we will have to handle this
differently.