[np-49257] refactor: Use persisted affiliations for access control (#500)
Ticket: https://sikt.atlassian.net/browse/NP-49257

This is initial refactoring to simplify work on the issue and not a
solution for it.
The main change is that we now use persisted top-level affiliations of
NVI Creators in access control checks, instead of fetching this data for
every request. This should also improve performance.