[np-49257] refactor: Replace access control logic (#502)
Ticket: https://sikt.atlassian.net/browse/NP-49257

This is (mostly) refactoring to structure the access control logic using
the same pattern as `nva-publication-api`. This is necessary to support
new "allowed operations", as the old system had the logic spread across
multiple levels.

Changes:
- Adds skeleton code for access control based on sets of explicit
grant/deny rules. An explicit deny will take precedence and override any
grants.
- `UserInstance`: Wrapper class for relevant metadata about a user,
extracted from a request to the REST API.
- `Strategy` files: Rules for access control and relevant helper methods
- `CandidatePermissions`: Determines allowed operations for a specific
`UserInstance` on a `Candidate` object.
- `CandidateResponseFactory`: Builder for a `CandidateDto` that manages
the user-specific metadata in the response. This is mostly code moved
out of the `Candidate` class.
- Rewrites some tests to use more narrow asserts on relevant fields
instead of recreating the entire expected response object.
- Rewrites other tests to pass a `UserInstance` along where a
`CandidateDto` response is needed.
- Uses the same access control code for
`Candidate.updateApprovalStatus`, so that the old access control
validator (`CandidateUpdateValidator`) can be deleted.