refactor: Restructure search tests for maintainability (#524)
Ticket: https://sikt.atlassian.net/browse/NP-49374

This is an initial refactoring to make it possible to write tests for
the planned changes.

Changes:
- Splits the existing `SearchNviCandidatesHandler` in three:
- Adds a base class that the other two test suites extend
(`SearchNviCandidatesHandlerTestBase`)
- The existing `SearchNviCandidatesHandlerTest` uses a mocked
OpenSearchClient (as before) and some static data, but with test setup
and assertions simplified. Hamcrest is removed and replaced with
AssertJ.
- `SearchNviCandidatesHandlerIntegrationTest` uses a TestContainer for
OpenSearch, making the tests integration tests. This is set up with
nested subclasses for different groups of tests (nearly empty
placeholders for now).
- Extracts reusable utility methods from `OpenSearchClientTest` and
`SearchNviCandidatesHandlerTest` to separate classes.