[np-49232] feat: Update search aggregation (#534)
Ticket: https://sikt.atlassian.net/browse/NP-49430

The core change here is that `ApprovalStatus.NEW` and
`ApprovalStatus.PENDING` are merged in the search aggregation, matching
the new query parameter changing. The PR includes some incidental
refactoring, but I've tried to limit it.

Note that the old aggregations are not removed yet, to give the
front-end time to adjust.

Changes:
- feat: Update aggregation filters to merge `new` and `pending` status
- refactor: Simplify query building by removing redundant nesting
- refactor: Simplify test setup for approval status tests
- refactor: Clean-up various linter warnings