feat: Add logging and test case for query complexity (#535)
This adds logging and testing to give us insight into the complexity of
the OpenSearch queries we generate.

Changes:

- Bumps `nva-commons`
- Add `TRACE` level logging of the full query string (JSON)
- Add `DEBUG` level logging of estimated query size
- Add `WARNING` level logging of unexpectedly complex queries. The
cut-off here is currently at 150. For comparison, current queries are
around ~300 and a "normal" query should probably be in the range 40-80.
- Set the environment variable `LOG_LEVEL` in all Gradle sub-projects.
Note that this must be set here to apply during tests, as the logging
framework does not read from the fake environment we otherwise use.
- Add parameter in `template.yaml` to set log level per environment,
with `debug` as default.

---------

Co-authored-by: Bruno Saxoni <789709+brinxmat@users.noreply.github.com>