[np-49493] fix: Handle indirect affiliations for access control (#542)
Ticket: https://sikt.atlassian.net/browse/NP-49493

This fixes a bug in access control that failed to grant access for
top-level NVI curators in cases where the affiliation was not a direct
child of the top-level organization.
This should not change the data that is persisted and therefore does not
require re-indexing, as the "workaround" is parsing the stored
organization hierarchy differently.


Changes:
- fix: Handle indirect affiliations for access control
- build: Disable Gradle configuration cache due to bugs in `sam build`