[np-49778] feat: Add prototype transaction locking (#552)
Ticket: https://sikt.atlassian.net/browse/NP-49778

This is a proof-of-concept of how we can handle concurrent writes that
may conflict (leading to lost updates and/or write skew) by using
optimistic locking and auto-incremented document versions.

Main changes:

- build: Bump `nva-commons` to `2.3.15`
- feat: Optimistic locking and concurrency control using features
built-in to DynamoDB. In short, a `revision` number is added to each
record and auto-incremented on each write. Writes include the last seen
revision number in each transaction which DynamoDB verifies. In
scenarios with concurrent writes based on the same data read, the first
write wins and the last write aborts.
- feat: Add auto-generated timestamp (`lastWrittenAt`) to
Candidate/Approval database model, set by DynamoDB on each write
operation. This is intended for debugging/auditing, but can also be used
for incremental re-indexing later if supported by an index.
- refactor: Change return types of some update methods to `void` to
force callers to re-fetch newest version from database after each
update.



Updating candidate and/or approvals must now be done as a single
transaction:

- `updateCandidate(candidate)`: `put` an entire candidate and fail on
mismatched`revision` in the database
- `updateCandidateAndApprovals( candidateDao,approvalsToUpdate,
approvalsToDelete)`: `PUT` a candidate, `PUT` for each approval to keep,
and `DELETE` for each approval to remove. Fails if any item has a
mismatched `revision` in the database.
- `updateApproval(candidate,approval)`: `PUT` a single approval with a
**condition** only on the expected `revision` of the parent `candidate`.
This fails if the candidate or approval has been updated concurrently,
but not on updates to other approvals for the same candidate.

Future work:

- This implementation does not handle large (>= 100) numbers of
top-level NVI organizations. This is because we wrap the writes in a
single transaction and this has an upper limit on both record size (400
KB?) and number of records (100?). For these cases, we should be able to
split the writing of approvals into batches, with a condition on the
`revision` of the candidate. I've added disabled (failing) test cases
for these scenarios.
- NOTE: Due to a new bug in expansion, this means the new implementation
will fail for many documents. This should not be deployed to `prod`
before at least one of these issues is resolved.
- The `version` field (random UUID) can be replaced entirely by the new
fields added. Actual data version is indicated by `revision`, while a
bump to `lastWrittenAt` would trigger re-indexing.
- Extend the implementation to cover all database writes (for Note,
Period etc) in the project
- Use strongly consistent reads in all read operations