fix: Normalize points on object creation (#557)
Adds canonical constructors to domain model records that normalize the
scale of `BigDecimal` values as the objects are created.

This may be necessary due to earlier changes that removed the custom
`equals() / hashcode()` methods. It also allows us to remove normalizing
logic from other layers and only handle it on object instantiation.