[np-50007] feat: Separate domain and database layers (#554)
Ticket: https://sikt.atlassian.net/browse/NP-50007

This is a significant architectural refactoring that separates domain
models from database persistence logic by introducing a service layer.
This is a necessary change in order to make it possible to properly fix
bugs related to concurrency and race conditions, as we need to
centralize and coordinate the writing of related data using
transactions.

Major changes:
- Converts domain models to immutable data classes and extracts any
database-related logic to the new service layer
  - Removed database repository dependencies from domain models
  - Introduced `CandidateService` for candidate lifecycle management
- Introduced `ApprovalService` for approval updates (status and
assignee)
  - Introduced `NoteService` for note creation and deletion
  - Extended the existing `NviPeriodService` for period management
  - Services coordinate between domain models and repositories
- Restructures repository classes and write operations to avoid
concurrency issues:
- Adds an async method to `CandidateRepository` to fetch all data
related to a candidate aggregate
- Uses consistent reads in all operations that support it (i.e. not GSI
queries)
- Makes methods that write data return `void` to force consumers to
re-fetch updated data
- Adds new method for fetching all related items (candidate, approvals,
notes) in a single, strongly consistent transaction
- Limits the direct use of repositories and `Dao` classes as much as
possible by adding/improving methods in domain models to convert to/from
`Dao` representations. It should be possible convert between the two
representations without loss of data.
 - Extended and improved data validation in domain models
 
 Future work:
- This does not fully introduce optimistic locking. The use of
`revision` and transaction checking should be extended to all the `Dao`
implementations.
- Transactional queries must be rewritten to handle large writes by
splitting them in relevant batches. Unlikely to be a problem for now,
but there is a limit to how many items a transaction can include.