[np-50246] Add tests for FetchInstitutionStatusAggregationHandler (#567)
Ticket: https://sikt.atlassian.net/browse/NP-50246

Adds integration tests for `FetchInstitutionStatusAggregationHandler`
that asserts the current (incorrect) behavior for a handful of basic
test cases. This is done with (hopefully temporary) static test data
because the endpoint currently returns handcrafted, artisanal "strings
that are hopefully valid JSON", which is not easily testable.

Apologies for the large PR, but it's mostly test data and helper methods
to set up predictable data in the index.