validating academic chapter as non valid (#579)
Validating AcademicChapter as non valid when it is a part of
AcademicMonograph or AcademicCommentary.
- Updated existing sparql for extraction of InstanceType because:
```
There are several UNION blocks referencing ?reference without properly binding it to the main publication.
This causes them to match any reference in the entire graph, including nested ones.
I bind ?reference to the main publication's reference in each UNION block.
```