[np-50294] fix: Remove period locking (#584)
Ticket: https://sikt.atlassian.net/browse/NP-50294

Removes locking based on `NviPeriod` from `CandidateRepository` to avoid
transaction errors, as described in the ticket.

Validation of period status is still done in the `Candidate` and
`EvaluatorService` classes, based on data read at the start of each
invokation (at the same time as the candidate data). Removing the
condition from the write transaction is safe because the effective
window is so small (a few seconds per year) and planned changes to
`NviPeriod` will limit the consequences.