[np-50446] fix: Avoid NPE in PublicationDto (#589)
Ticket: https://sikt.atlassian.net/browse/NP-50446

Fixes nullability issues in `PublicationDto` that caused errors in data
migration. This could happen if a publication was no longer valid as an
NVI candidate (i.e. no contributors), but was already reported and
couldn't be removed/updated. The migration process adds missing data
based on the current state of the publication, but shouldn't fail even
if the publication is no longer completely valid as a candidate.

Changes:
- Sets a default value (empty list) for collections to avoid
null-pointer exceptions.
- Adds check for empty list to `validate()` method
- Refactors test helpers to use valid data